WorldGen: Make tree generation fully deterministic
There were a few `Collections.shuffle()` calls used to generate some of the
randomness for world generation. Sadly, this does not respect the `RandomSource`
used by Minecraft.

This PR adds a new WorldgenHelper that shuffles Collections by using a given
RandomSource and then uses it for tree and ore generation. This ensures that the
same world seeds generate the same ore veins and trees.